Hydrological pre-processing workflow for digital elevation models

Hydrological pre-processing workflow for digital elevation models

Usage

flow_accumlation_wb(
  dem_file_path,
  ModelFolder,
  watershed_shape_path = NA,
  max_dist = 1000,
  stream_threshold = 1200,
  carve = 1,
  overwrite = T
)

Arguments

dem_file_path

path to digital elevation file

ModelFolder

Output path to save model outputs

watershed_shape_path

Optional shapefile to clip the resulting results

max_dist

numeric value in meters to breach the digital elevation model

stream_threshold

Flow accumulation number (cells) to create channels

carve

numeric. Number of meters to carve the channel network down by.

overwrite

T/F. If TRUE, the model will overwrite previous files in Model Folder

Value

SpatRaster of the modified digital elevation model

Examples

ModelFolder <- tempdir() # Create a temporary directory to save files into